fix PokemonFragment so that the pokemon's evolution chain is correctly matched with the correct pokemon
After removing some unneccessary for loops and fixing the indexes, I got
pokedex to show me to correct evolution chain for each pokemon. This
took me a lot of debugging but I eventually got it working. Still have
to fix the error at the end because I run out of pokemon.